Fixes on Line Tracking plus some code structure changes
- Line Tracking now use previous tangential direction for a point that has a (0,0) tan value to avoid a infinite loop.
- Moved CGPoint+SPShapener and some extensions in SPLineGroupVectorizor into Direction2D.swift.
- And some more code movements.

Tested Line Tracking on a bigger image that it wasn't able to handle and the result is quite good.

But I guess there is something wrong with my flood method.

Junction point detection is not good, which causes smoothDirection to be wrong. So next, I am going to work on that applying my own method.